fix(api): Add missing go.sum and resolve build issues
- Generate missing api/go.sum file (required for CI/CD)
- Fix incorrect import path in internal/handlers/nodes.go
- Fix pointer type errors in activity tracker
- Add missing Repository type to models/plugin.go
- Add missing Username field to models/UserQuota
- Remove duplicate ErrorResponse/SuccessResponse types, create shared types.go
- Remove unused imports in cache/middleware.go and handlers/catalog.go

This fixes the primary cause of workflow failures where go.mod could not
download dependencies due to missing checksums in go.sum.

Related to security scan failures in GitHub Actions run #19377892516.
